requirements.txt
pylyrics
Google-Search-API
